ref: implement new factory
